requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Pillow==10.0.0
rembg[cpu]==2.0.50
gunicorn==21.2.0